// client/src/interfaces/IAddress.ts
export interface IAddress {
    id?: number;              
    fullname: string;        
    phone: string;           
    province: string;      
    district: string;       
    ward: string;           
    address: string;       
    user_id: number;      
    created_at: string; 
    updated_at: string;
}